requirements.txt
asgiref==3.2.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-import-export==2.0.1
et-xmlfile==1.0.1
freeze==1.0.10
jdcal==1.4.1
MarkupPy==1.14
mysql==0.0.2
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
pytz==2019.3
PyYAML==5.3
reportlab==3.5.34
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
